+++ packages/flutter_chat_ui/lib/src/chat_animated_list/chat_animated_list_reversed.dart
@@ -95,7 +95,7 @@ class ChatAnimatedListReversedState extends State<ChatAnimatedListReversed>
     _observerController = SliverObserverController(
       controller: _scrollController,
     )..cacheJumpIndexOffset = false;
-    // TODO: Add assert for messages having same id
+
     _oldList = List.from(_chatController.messages);
     _operationsSubscription = _chatController.operationsStream.listen((event) {
       switch (event.type) {